+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- `s.players()` can now accept an array of configs and will return an array of Players. Passing an object of configs has the same effect as previously
+- There is now an `Instance` object which represents an instance inside a conductor. It can be accessed like so:
+```javascript
+const instance = player.instances['instanceId']
+ 
+// the following two lines are equivalent
+await instance.call('zome', 'fn', params)
+await player.call('instanceId', 'zome', 'fn', params) 
+```
+
 ## [0.1.1] - 2019-10-08
 
 ### Added